fix parsing uischema registry entries
