+++ packages/jsonforms-vuetify-renderers/src/util/index.ts
@@ -1,1 +1,2 @@
 export * from './composition';
+export * from './util';
